refactor(progresses): dafault/user/seller 가입 화면 파일 나누기
