+++ src/pages/Signup/progresses/UserProgress.tsx
@@ -1,0 +1,103 @@
+import { Text } from '@chakra-ui/react';
+import { useState } from 'react';
+
+import HorizontalLine from '@/components/styles/HorizontalLine';
+import MembershipAgreement from '../MembershipClauses';
+import ProgressBar from '../ProgressBar';
+import { InputItem, ProgressBox, StyledInput } from './styles';
+
+const UserProgress = () => {
+  const [birthDate, setBirthDate] = useState<string>('');
+  const [phone, setPhone] = useState<string>('');
+  const [email, setEmail] = useState<string>('');
+  const [isBirthDateValid, setIsBirthDateValid] = useState<boolean>(true);
+  const [isPhoneValid, setIsPhoneValid] = useState<boolean>(true);
+  const [isEmailValid, setIsEmailValid] = useState<boolean>(true);
+
+  // 값 업데이트 및 유효성 검사
+  const handleBirthDateChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setBirthDate(e.target.value);
+
+    const birthDateRegex = /^\d{4}-\d{2}-\d{2}$/;
+    setIsBirthDateValid(birthDateRegex.test(e.target.value));
+  };
+
+  const handlePhoneChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setPhone(e.target.value);
+
+    const phoneRegex = /^\d{3}-\d{4}-\d{4}$/;
+    setIsPhoneValid(phoneRegex.test(e.target.value));
+  };
+
+  const handleEmailChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setEmail(e.target.value);
+
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    setIsEmailValid(emailRegex.test(e.target.value));
+  };
+
+  return (
+    <ProgressBox>
+      <ProgressBar percentage={100} />
+      <p className="progress-guidance">
+        000 님, 반가워요.
+        <br />
+        회원 정보를 입력해주세요.
+      </p>
+      <form className="progress-form">
+        <InputItem>
+          <p className="input-label">생년월일 *</p>
+          <StyledInput valid={isBirthDateValid}>
+            <input
+              type="date"
+              className="input-element"
+              value={birthDate}
+              onChange={handleBirthDateChange}
+            />
+            {!isBirthDateValid && <p className="input-validation">생년월일을 다시 확인해주세요.</p>}
+          </StyledInput>
+        </InputItem>
+        <InputItem>
+          <p className="input-label">휴대 전화 *</p>
+          <StyledInput valid={isPhoneValid}>
+            <input
+              type="number"
+              className="input-element"
+              placeholder="000-0000-0000"
+              value={phone}
+              onChange={handlePhoneChange}
+            />
+            {!isPhoneValid && <p className="input-validation">휴대 전화를 다시 확인해주세요.</p>}
+          </StyledInput>
+        </InputItem>
+        <InputItem>
+          <p className="input-label">이메일 *</p>
+          <StyledInput valid={isEmailValid}>
+            <input
+              type="email"
+              className="input-element"
+              placeholder="abc@1618.com"
+              value={email}
+              onChange={handleEmailChange}
+            />
+            {!isEmailValid && <p className="input-validation">이메일을 다시 확인해주세요.</p>}
+          </StyledInput>
+        </InputItem>
+        <InputItem>
+          <p className="input-label">관심사</p>
+          <StyledInput valid={true}>
+            <div className="input-element">
+              <Text color="var(--color-gray-dk)">
+                관심사를 등록하고 관련 작품을 추천받아보세요.
+              </Text>
+            </div>
+          </StyledInput>
+        </InputItem>
+        <HorizontalLine />
+        <MembershipAgreement />
+      </form>
+    </ProgressBox>
+  );
+};
+
+export default UserProgress;
